Fix broken sharelink button
Yasqe has switched from using GET params to using a fragment (which is
structured like GET params ...). So the createShareLink function now creates a
fragment, and the getUrlParams ignores  _all_ GET params when a hash character
is found in the url. The fix is adding the endpoint param to the shareLink
object explictly, so it becomes part of the fragment.

Fix execution with query param (has to be asyncronous)